requirements.txt
openai==1.50.2
python-dotenv==1.0.1
schedule==1.2.2
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
matplotlib==3.5.3
wordcloud==1.9.2
pandas==1.5.3
scikit-learn==1.2.2
flask==2.3.2
redis==4.5.5
rq==1.15.1
numpy==1.23.5